refclock: fix compiler warning in SOCK error message

+++ refclock_sock.c
@@ -64,7 +64,7 @@ static void read_sample(void *anything)
 
   if (s != sizeof (sample)) {
     LOG(LOGS_WARN, LOGF_Refclock, "Unexpected length of SOCK sample : %d != %ld",
-        s, sizeof (sample));
+        s, (long)sizeof (sample));
     return;
   }
